// src/components/Client/InterfaceStructures.ts
export interface IClient {
    customerID: number;
    rcpCustomer: null;
    customerCode: null;
    firstName: null;
    lastName: null;
    birthday: null;
    contactPhone: string;
    workPhone: null;
    email: string;
    title: null;
    address: null;
    notes: null;
    customerName: string;
    city: null;
    state: null;
    zip: null;
    totalSpentByYear: number;
    country: null;
    imageFileName: null;
    joinDate: null;
    passwordLoginWeb: null;
    visitCount: null;
    fristVist: null;
    lastVisit: Date;
    favouritePolish: null;
    favouriteTechs1: string;
    favouriteTechs2: string;
    favouriteTechs3: string;
    favouriteTech: string;
    notesApp: null;
    coupon: string;
    rewardsPoint: number;
    rating: number;
    memberStatus: string;
    isKid: null;
    isChild: boolean;
    ratingDate: null;
    rewardsMember: null;
    isClientVerify: null;
    memberID: null;
    verification: null;
    isVerifyPhoneWithMango: null;
    isChangePhoneWhenReward: null;
    isChangePhone: boolean;
    visitCountByYear: null;
    isBlackList: boolean;
    isDeleted: null;
    atRisk: null;
    isVip: null;
    totalAmount: null;
    customerType: string;
  }